+++ Sources/InfinitusUI/T3/Lucide.generated.swift
@@ -105,7 +105,10 @@ public enum Lucide: String, CaseIterable, Sendable {
     case smartphone = "smartphone"
     case smilePlus = "smile-plus"
     case sparkles = "sparkles"
+    case square = "square"
     case squarePen = "square-pen"
+    case squareSplitHorizontal = "square-split-horizontal"
+    case squareSplitVertical = "square-split-vertical"
     case star = "star"
     case sun = "sun"
     case tag = "tag"
@@ -232,7 +235,10 @@ public enum Lucide: String, CaseIterable, Sendable {
         case .smartphone: return ["M7.0 2.0h10.0a2.0 2.0 0 0 1 2.0 2.0v16.0a2.0 2.0 0 0 1 -2.0 2.0h-10.0a2.0 2.0 0 0 1 -2.0 -2.0v-16.0a2.0 2.0 0 0 1 2.0 -2.0z", "M12 18h.01"]
         case .smilePlus: return ["M22 11v1a10 10 0 1 1-9-10", "M8 14s1.5 2 4 2 4-2 4-2", "M9.0 9.0L9.01 9.0", "M15.0 9.0L15.01 9.0", "M16 5h6", "M19 2v6"]
         case .sparkles: return ["M11.017 2.814a1 1 0 0 1 1.966 0l1.051 5.558a2 2 0 0 0 1.594 1.594l5.558 1.051a1 1 0 0 1 0 1.966l-5.558 1.051a2 2 0 0 0-1.594 1.594l-1.051 5.558a1 1 0 0 1-1.966 0l-1.051-5.558a2 2 0 0 0-1.594-1.594l-5.558-1.051a1 1 0 0 1 0-1.966l5.558-1.051a2 2 0 0 0 1.594-1.594z", "M20 2v4", "M22 4h-4", "M2.0 20.0a2.0 2.0 0 1 0 4.0 0a2.0 2.0 0 1 0 -4.0 0"]
+        case .square: return ["M5.0 3.0h14.0a2.0 2.0 0 0 1 2.0 2.0v14.0a2.0 2.0 0 0 1 -2.0 2.0h-14.0a2.0 2.0 0 0 1 -2.0 -2.0v-14.0a2.0 2.0 0 0 1 2.0 -2.0z"]
         case .squarePen: return ["M12 3H5a2 2 0 0 0-2 2v14a2 2 0 0 0 2 2h14a2 2 0 0 0 2-2v-7", "M18.375 2.625a1 1 0 0 1 3 3l-9.013 9.014a2 2 0 0 1-.853.505l-2.873.84a.5.5 0 0 1-.62-.62l.84-2.873a2 2 0 0 1 .506-.852z"]
+        case .squareSplitHorizontal: return ["M8 19H5c-1 0-2-1-2-2V7c0-1 1-2 2-2h3", "M16 5h3c1 0 2 1 2 2v10c0 1-1 2-2 2h-3", "M12.0 4.0L12.0 20.0"]
+        case .squareSplitVertical: return ["M5 8V5c0-1 1-2 2-2h10c1 0 2 1 2 2v3", "M19 16v3c0 1-1 2-2 2H7c-1 0-2-1-2-2v-3", "M4.0 12.0L20.0 12.0"]
         case .star: return ["M11.525 2.295a.53.53 0 0 1 .95 0l2.31 4.679a2.123 2.123 0 0 0 1.595 1.16l5.166.756a.53.53 0 0 1 .294.904l-3.736 3.638a2.123 2.123 0 0 0-.611 1.878l.882 5.14a.53.53 0 0 1-.771.56l-4.618-2.428a2.122 2.122 0 0 0-1.973 0L6.396 21.01a.53.53 0 0 1-.77-.56l.881-5.139a2.122 2.122 0 0 0-.611-1.879L2.16 9.795a.53.53 0 0 1 .294-.906l5.165-.755a2.122 2.122 0 0 0 1.597-1.16z"]
         case .sun: return ["M8.0 12.0a4.0 4.0 0 1 0 8.0 0a4.0 4.0 0 1 0 -8.0 0", "M12 2v2", "M12 20v2", "m4.93 4.93 1.41 1.41", "m17.66 17.66 1.41 1.41", "M2 12h2", "M20 12h2", "m6.34 17.66-1.41 1.41", "m19.07 4.93-1.41 1.41"]
         case .tag: return ["M12.586 2.586A2 2 0 0 0 11.172 2H4a2 2 0 0 0-2 2v7.172a2 2 0 0 0 .586 1.414l8.704 8.704a2.426 2.426 0 0 0 3.42 0l6.58-6.58a2.426 2.426 0 0 0 0-3.42z", "M7.0 7.5a0.5 0.5 0 1 0 1.0 0a0.5 0.5 0 1 0 -1.0 0"]